Move code into own documentation provider
