+++ packages/langium/src/documentation/index.ts
@@ -1,0 +1,8 @@
+/******************************************************************************
+ * Copyright 2023 TypeFox GmbH
+ * This program and the accompanying materials are made available under the
+ * terms of the MIT License, which is available in the project root.
+ ******************************************************************************/
+
+export * from './documentation-provider';
+export * from './jsdoc';
