+++ packages/langium/src/index.ts
@@ -6,6 +6,7 @@
 
 export * from './default-module';
 export * from './dependency-injection';
+export * from './documentation';
 export * from './generator/generator-node';
 export * from './generator/template-node';
 export { expandToString, expandToStringWithNL, normalizeEOL } from './generator/template-string';
@@ -31,7 +32,6 @@ export * from './utils/ast-util';
 export * from './utils/collections';
 export * from './utils/cst-util';
 export * from './utils/grammar-util';
-export * from './utils/jsdoc-util';
 export * from './utils/promise-util';
 export * from './utils/uri-util';
 export * from './utils/regex-util';
