+++ packages/langium/src/services.ts
@@ -54,6 +54,7 @@ import type { CallHierarchyProvider } from './lsp/call-hierarchy-provider';
 import type { DocumentLinkProvider } from './lsp/document-link-provider';
 import type { CodeLensProvider } from './lsp/code-lens-provider';
 import type { DeclarationProvider } from './lsp/declaration-provider';
+import type { DocumentationProvider } from './documentation/documentation-provider';
 
 /**
  * The services generated by `langium-cli` for a single language. These are derived from the
@@ -101,6 +102,9 @@ export type LangiumDefaultServices = {
         CompletionParser: LangiumCompletionParser
         TokenBuilder: TokenBuilder
         Lexer: Lexer
+    }
+    documentation: {
+        DocumentationProvider: DocumentationProvider
     }
     references: {
         Linker: Linker
